Extract useBarListeners from App.tsx
The biggest single refactor of the series. Pull all 5 onSnapshot
subscriptions (per-bar user profile, bar config, requests, who's-on,
notices) plus the FCM-token-arrives-→-write tokens/{uid}
auto-clock-in side effect into useBarListeners({ user, barId,
fcmToken, globalNtfyTopic, setSearchParams }).

The two internal effects are kept separate on purpose:
- listener effect deps [user, barId, setSearchParams] — when any of
  these change, all 5 listeners tear down and re-subscribe.
- auto-clock-in effect deps [user, barId, fcmToken] — fires once
  the token arrives without disturbing the listeners.

globalNtfyTopic is captured via a ref inside the listener effect so
new topic values are picked up on the next per-bar user snapshot
without re-subscribing. This preserves the ListenerRedundancy
benchmark's "<12 onSnapshot calls during FCM token update" budget
(was 16 calls before adding the ref; same as pre-extraction).

Hook returns:
- Read-only: userRole, userStatus, displayName, ntfyTopic, barName,
  buttons, beerInventory, wells, hiddenButtonIds, buttonUsage,
  customOrders, requests, allUsers, notices
- Setters that escape because they're used for optimistic mutation
  elsewhere in App: setNotificationPreferences, setNtfyTopic,
  setBeerInventory, setWells, setHiddenButtonIds, setCustomOrders

App.tsx drops 11 firebase/firestore imports plus the Bar/Request/
Notice/DEFAULT_BUTTONS imports that only the listener block needed.

App.tsx down from 1454 to ~1310 lines. 48/48 tests pass; tsc clean;
vite build succeeds.
